[CP-1519] show added contact properly after adding next contact
